Conflicts:
	geonode/context_processors.py
	geonode/layers/models.py
	geonode/layers/templates/layers/layer_detail.html
	geonode/layers/templates/layers/layer_leaflet_map.html
	geonode/layers/views.py
	geonode/services/fixtures/initial_data.json
	geonode/services/models.py
	geonode/services/templates/services/service_register.html
	geonode/services/views.py
	geonode/settings.py
	geonode/urls.py

+++ geonode/layers/models.py
@@ -78,7 +78,7 @@ class Layer(ResourceBase):
     store = models.CharField(max_length=128)
     storeType = models.CharField(max_length=128)
     name = models.CharField(max_length=128)
-    typename = models.CharField(max_length=128, unique=True, null=True, blank=True)
+    typename = models.CharField(max_length=128, null=True, blank=True)
 
     default_style = models.ForeignKey(Style, related_name='layer_default_style', null=True, blank=True)
     styles = models.ManyToManyField(Style, related_name='layer_styles')
@@ -86,6 +86,8 @@ class Layer(ResourceBase):
     charset = models.CharField(max_length=255, default='UTF-8')
 
     upload_session = models.ForeignKey('UploadSession', blank=True, null=True)
+
+    service = models.ForeignKey('services.Service', null=True, blank=True, related_name='layer_set')
 
     def is_vector(self):
         return self.storeType == 'dataStore'
@@ -123,10 +125,24 @@ class Layer(ResourceBase):
     @property
     def ows_url(self):
         if self.storeType == "remoteStore":
-            from geonode.services.models import ServiceLayer
-            return ServiceLayer.objects.filter(layer__id=self.id)[0].service.base_url
+            return self.service.base_url
         else:
             return settings.OGC_SERVER['default']['LOCATION'] + "wms"
+
+
+    @property
+    def ptype(self):
+        if self.storeType == "remoteStore":
+            return self.service.ptype
+        else:
+            return "gxp_wmscsource"
+
+    @property
+    def service_typename(self):
+        if self.storeType == "remoteStore":
+            return "%s:%s" % (self.service.name, self.typename)
+        else:
+            return self.typename
 
     def get_base_file(self):
         """Get the shp or geotiff file for this layer.
@@ -150,7 +166,7 @@ class Layer(ResourceBase):
 
 
     def get_absolute_url(self):
-        return reverse('layer_detail', args=(self.typename,))
+        return reverse('layer_detail', args=(self.service_typename,))
 
 
     def attribute_config(self):
@@ -166,12 +182,24 @@ class Layer(ResourceBase):
 
     def __str__(self):
         if self.typename is not None:
-            return "%s Layer" % self.typename.encode('utf-8')
+            return "%s Layer" % self.service_typename.encode('utf-8')
         elif self.name is not None:
             return "%s Layer" % self.name
         else:
             return "Unamed Layer"
 
+
+    class Meta:
+        # custom permissions,
+        # change and delete are standard in django
+        permissions = (('view_layer', 'Can view'),
+                       ('change_layer_permissions', "Can change permissions"), )
+
+    # Permission Level Constants
+    # LEVEL_NONE inherited
+    LEVEL_READ  = 'layer_readonly'
+    LEVEL_WRITE = 'layer_readwrite'
+    LEVEL_ADMIN = 'layer_admin'
 
     def maps(self):
         from geonode.maps.models import MapLayer
@@ -180,9 +208,6 @@ class Layer(ResourceBase):
     @property
     def class_name(self):
         return self.__class__.__name__
-
-    class Meta(ResourceBase.Meta):
-        pass
 
 class Layer_Styles(models.Model):
     layer = models.ForeignKey(Layer)
@@ -254,7 +279,6 @@ class Attribute(models.Model):
     def unique_values_as_list(self):
         return self.unique_values.split(',')
 
-
 def pre_save_layer(instance, sender, **kwargs):
     if kwargs.get('raw', False):
         instance.owner = instance.resourcebase_ptr.owner
@@ -317,6 +341,8 @@ def pre_delete_layer(instance, sender, **kwargs):
     Remove any associated style to the layer, if it is not used by other layers.
     Default style will be deleted in post_delete_layer
     """
+    if instance.service:
+        return
     logger.debug("Going to delete the styles associated for [%s]", instance.typename.encode('utf-8'))
     ct = ContentType.objects.get_for_model(instance)
     OverallRating.objects.filter(content_type = ct, object_id = instance.id).delete()
@@ -333,7 +359,10 @@ def post_delete_layer(instance, sender, **kwargs):
     """
     from geonode.maps.models import MapLayer
     logger.debug("Going to delete associated maplayers for [%s]", instance.typename.encode('utf-8'))
-    MapLayer.objects.filter(name=instance.typename).delete()
+    MapLayer.objects.filter(name=instance.typename,ows_url=instance.ows_url).delete()
+
+    if instance.service:
+        return
     logger.debug("Going to delete the default style for [%s]", instance.typename.encode('utf-8'))
 
     if instance.default_style and Layer.objects.filter(default_style__id=instance.default_style.id).count() == 0:
